# docs/swagger.yaml
basePath: /
definitions:
  contracts.ErrPayload:
    properties:
      code:
        type: string
      description:
        type: string
    type: object
  contracts.ErrResponse:
    properties:
      error:
        $ref: '#/definitions/contracts.ErrPayload'
    type: object
  contracts.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_rows:
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Exercise:
    properties:
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      trainingPlanID:
        type: integer
      trainingPlanVersion:
        type: integer
    type: object
...
  models.Review:
    properties:
      comment:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      score:
        type: integer
      trainingPlanID:
        type: integer
      trainingPlanVersion:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: string
    type: object
  models.Tag:
    properties:
      name:
        type: string
    type: object
  models.TrainingPlan:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      difficulty:
        type: string
      disabled:
        type: boolean
      duration:
        type: integer
      exercises:
        items:
          $ref: '#/definitions/models.Exercise'
        type: array
      favoritesCount:
        type: integer
      id:
        type: integer
      meanScore:
        type: number
      name:
        type: string
      pictureUrl:
        type: string
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      sessionsCount:
        type: integer
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      trainerID:
        type: string
      version:
        type: integer
    type: object
...
  models.User:
    properties:
      bornAt:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      displayName:
        type: string
      height:
        type: integer
      id:
        type: string
      interests:
        items:
          $ref: '#/definitions/models.UserInterest'
        type: array
      isMale:
        type: boolean
      isVerifiedTrainer:
        type: boolean
      mainLocation:
        type: string
      nickname:
        type: string
      pictureUrl:
        type: string
      weight:
        type: integer
    type: object
  models.UserInterest:
    properties:
      name:
        type: string
    type: object
  reviews.CreateReviewRequest:
    properties:
      comment:
        type: string
      score:
        type: integer
      trainingPlanID:
        type: integer
      user_id:
        type: string
    required:
    - score
    - user_id
    type: object
  reviews.DeleteReviewRequest:
    properties:
      reviewID:
        type: integer
      trainingPlanID:
        type: integer
      user_id:
        type: string
    required:
    - user_id
    type: object
  reviews.GetReviewsResponse:
    properties:
      pagination:
        $ref: '#/definitions/contracts.Pagination'
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
    type: object
  reviews.UpdateReviewRequest:
    properties:
      comment:
        type: string
      reviewID:
        type: integer
      score:
        type: integer
      trainingPlanID:
        type: integer
      user_id:
        type: string
    required:
    - user_id
    type: object
...
  trainings.DeleteTrainingRequest:
    properties:
      trainer_id:
        type: string
      trainingPlanID:
        type: integer
    required:
    - trainer_id
    type: object
  trainings.ExerciseRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  trainings.UpdateTrainingRequest:
    properties:
      description:
        type: string
      difficulty:
        type: string
      duration:
        type: integer
      exercises:
        items:
          $ref: '#/definitions/trainings.ExerciseRequest'
        type: array
      id:
        type: integer
      name:
        type: string
      tags:
        items:
          type: string
        type: array
      trainer_id:
        type: string
    required:
    - description
    - difficulty
    - duration
    - exercises
    - name
    - trainer_id
    type: object
...
host: fiufit-trainings.fly.dev
info:
  contact: {}
  description: Fiufit's Trainings service documentation. This service manages training
    plans and its sessions, goals and reviews, etc.
  title: Fiufit Trainings API
  version: dev
paths:
...
  /{version}/trainings/{trainingID}/enable:
    post:
      consumes:
      - application/json
      description: Enables a training plan that had been disabled previously given
        its ID
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: Training plan ID
        in: path
        name: trainingID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Enables a training plan
      tags:
      - training_plans
...
  /{version}/trainings/{trainingID}/reviews:
    get:
      consumes:
      - application/json
      description: Get reviews of a training plan with pagination filtered by their
        score, comment or userID
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: Training Plan ID
        in: path
        name: trainingID
        required: true
        type: integer
      - description: minimum score of the review
        in: query
        name: min_score
        type: integer
      - description: maximum score of the review
        in: query
        name: max_score
        type: integer
      - description: comment of the review
        in: query
        name: comment
        type: string
      - description: userID of the review
        in: query
        name: user_id
        type: string
      - description: page number when getting with pagination
        in: query
        name: page
        type: integer
      - description: page size when getting with pagination
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            $ref: '#/definitions/reviews.GetReviewsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Get reviews by different query params with pagination
      tags:
      - reviews
    post:
      consumes:
      - application/json
      description: Creates a new review with a comment and a score for a given trainingPlanID
        and userID
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: Training Plan ID
        in: path
        name: trainingID
        required: true
        type: integer
      - description: Body params
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/reviews.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            $ref: '#/definitions/models.Review'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Creates a review for a training plan
      tags:
      - reviews
  /{version}/trainings/{trainingID}/reviews/{reviewID}:
    delete:
      consumes:
      - application/json
      description: Deletes a review for a given reviewID and trainingID given by the
        userID
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: Training Plan ID
        in: path
        name: trainingID
        required: true
        type: integer
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: integer
      - description: Body params
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/reviews.DeleteReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Deletes a review
      tags:
      - reviews
    get:
      consumes:
      - application/json
      description: Gets a review for a given reviewID
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: Training Plan ID
        in: path
        name: trainingID
        required: true
        type: integer
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            $ref: '#/definitions/models.Review'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Gets a review by ID
      tags:
      - reviews
    patch:
      consumes:
      - application/json
      description: Updates the comment and/or rating of a review given by a user for
        a training plan
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: Training Plan ID
        in: path
        name: trainingID
        required: true
        type: integer
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: integer
      - description: Body params
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/reviews.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Updates a review
      tags:
      - reviews
  /{version}/trainings/favorites:
    get:
      consumes:
      - application/json
      description: Gets favorite training plans for a user with pagination
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Gets favorite training plans for a user
      tags:
      - training_plans
  /{version}/trainings{training_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a training plan for a given trainingID and validates that
        the user has permissions to do so
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: Training ID
        in: path
        name: trainingID
        required: true
        type: integer
      - description: Body params
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/trainings.DeleteTrainingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Deletes a training plan
      tags:
      - training_plans
    put:
      consumes:
      - application/json
      description: Updates a training plan with a name, description, difficulty, duration,
        exercises and tags
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: Training ID
        in: path
        name: trainingID
        required: true
        type: integer
      - description: Body params
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/trainings.UpdateTrainingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            $ref: '#/definitions/models.TrainingPlan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Updates a training plan
      tags:
      - training_plans
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"